Added Elixir setup
Added a script `elixir.sh` and edited the .dot file to have a script to set up an elixir environment.

The `elixir.sh` script installs elixir and node through brew and then sets up the Hex package manager for elixir and installs phoenix.

To run this script add provide the argument 'elixir' when you run the .dots file.